+++ packages/create-rsbuild/template-vanilla-ts/src/index.ts
@@ -2,7 +2,7 @@ import './index.css';
 
 document.querySelector('#root')!.innerHTML = `
 <div class="content">
-  <h1>Rsbuild with React</h1>
+  <h1>Vanilla Rsbuild</h1>
   <p>Start building amazing things with Rsbuild.</p>
 </div>
 `;