+++ packages/doctor-core/src/rules/index.ts
@@ -1,0 +1,4 @@
+export * from './linter';
+export * from './rule';
+
+export { rules } from './rules';