+++ packages/shared/src/plugins/index.ts
@@ -1,4 +1,3 @@
-export { AutoSetRootFontSizePlugin } from './AutoSetRootFontSizePlugin';
 export { HtmlTagsPlugin } from './HtmlTagsPlugin';
 export type { HtmlTagsPluginOptions } from './HtmlTagsPlugin';
 export { HtmlCrossOriginPlugin } from './HtmlCrossOriginPlugin';