+++ packages/plugin-vue-jsx/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @rsbuild/plugin-vue-jsx
 
+## 0.0.21
+
+### Patch Changes
+
+- Updated dependencies [42ace29]
+- Updated dependencies [a919119]
+- Updated dependencies [0270c27]
+- Updated dependencies [dbdc3c5]
+- Updated dependencies [d691901]
+- Updated dependencies [a694a55]
+- Updated dependencies [8aa416e]
+- Updated dependencies [69262e9]
+- Updated dependencies [6abd810]
+- Updated dependencies [4092858]
+- Updated dependencies [648e847]
+  - @rsbuild/core@0.0.21
+
 ## 0.0.20
 
 ### Patch Changes